Make history dto in shared folder
