+++ client/src/components/content/abstract-content.ts
@@ -1,13 +1,5 @@
 import Component from '../component';
-
-export type History = {
-	id: number;
-	price: number;
-	content: string;
-	historyDate: Date;
-	category: string;
-	payment: string;
-};
+import { History } from '@shared/dto/history-dto';
 
 export abstract class AbstractContent extends Component {
 	// 새로운 기록을 받아서 렌더/업데이트
